+++ src/ViewHoldingsRecord.js
@@ -14,7 +14,6 @@ import queryString from 'query-string';
 import {
   Layer,
   Pane,
-  PaneMenu,
   Row,
   Col,
   Accordion,
@@ -22,7 +21,6 @@ import {
   AccordionStatus,
   ExpandAllButton,
   KeyValue,
-  PaneHeaderIconButton,
   MultiColumnList,
   Icon,
   Button,
@@ -295,7 +293,6 @@ class ViewHoldingsRecord extends React.Component {
       },
       referenceTables,
       okapi,
-      stripes,
     } = this.props;
 
     if (this.isAwaitingResource()) {
@@ -308,23 +305,6 @@ class ViewHoldingsRecord extends React.Component {
     const holdingsTemporaryLocation = holdingsRecord.temporaryLocationId ? temporaryLocation.records[0] : null;
     const itemCount = get(items, 'records.length', 0);
     const query = location.search ? queryString.parse(location.search) : {};
-
-    const detailMenu = stripes.hasPerm('ui-inventory.holdings.edit') && (
-      <PaneMenu>
-        <FormattedMessage id="ui-inventory.editHoldings">
-          {ariaLabel => (
-            <PaneHeaderIconButton
-              icon="edit"
-              id="clickable-edit-holdingsrecord"
-              style={{ visibility: !holdingsRecord ? 'hidden' : 'visible' }}
-              href={this.craftLayerUrl('editHoldingsRecord')}
-              onClick={this.onClickEditHoldingsRecord}
-              ariaLabel={ariaLabel}
-            />
-          )}
-        </FormattedMessage>
-      </PaneMenu>
-    );
 
     const confirmHoldingsRecordDeleteModalMessage = (
       <SafeHTMLMessage
@@ -541,7 +521,6 @@ class ViewHoldingsRecord extends React.Component {
                       />
                     </span>
                   }
-                  lastMenu={detailMenu}
                   dismissible
                   onClose={this.props.onCloseViewHoldingsRecord}
                   actionMenu={this.getPaneHeaderActionMenu}