+++ src/ViewInstance.js
@@ -14,11 +14,7 @@ import {
 } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import ReactRouterPropTypes from 'react-router-prop-types';
-import {
-  FormattedDate,
-  FormattedTime,
-  FormattedMessage,
-} from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import queryString from 'query-string';
 
@@ -30,7 +26,6 @@ import {
 } from '@folio/stripes/core';
 import {
   Pane,
-  PaneMenu,
   Row,
   Col,
   Button,
@@ -41,7 +36,6 @@ import {
   KeyValue,
   Layer,
   Layout,
-  PaneHeaderIconButton,
   Icon,
   Headline,
   MultiColumnList,
@@ -55,9 +49,10 @@ import {
   checkIfElementIsEmpty,
   convertArrayToBlocks,
   checkIfArrayIsEmpty,
-  staffOnlyFormatter,
+  getDateWithTime,
   getSortedNotes,
   marshalInstance,
+  staffOnlyFormatter,
   unmarshalInstance,
 } from './utils';
 import formatters from './referenceFormatters';
@@ -338,6 +333,43 @@ class ViewInstance extends React.Component {
             </Icon>
           </Button>
         </IfPermission>
+
+        {
+          isSourceMARC && (
+            <>
+              <IfPermission perm="ui-inventory.instance.view">
+                <Button
+                  id="clickable-view-source"
+                  buttonStyle="dropdownItem"
+                  disabled={!marcRecord}
+                  onClick={(e) => {
+                    onToggle();
+                    this.handleViewSource(e, instance);
+                  }}
+                >
+                  <Icon icon="document">
+                    <FormattedMessage id="ui-inventory.viewSource" />
+                  </Icon>
+                </Button>
+              </IfPermission>
+
+              <IfPermission perm="records-editor.records.item.put">
+                <Button
+                  id="edit-instance-marc"
+                  buttonStyle="dropdownItem"
+                  onClick={() => {
+                    onToggle();
+                    this.editInstanceMarc();
+                  }}
+                >
+                  <Icon icon="edit">
+                    <FormattedMessage id="ui-inventory.editInstanceMarc" />
+                  </Icon>
+                </Button>
+              </IfPermission>
+            </>
+          )
+        }
       </Fragment>
     );
   };
@@ -429,25 +461,6 @@ class ViewInstance extends React.Component {
       'Source': x => this.refLookup(referenceTables.instanceFormats, x.id).source || noValue,
     };
 
-    const detailMenu = (
-      <IfPermission perm="ui-inventory.instance.edit">
-        <PaneMenu>
-          <FormattedMessage id="ui-inventory.editInstance">
-            {ariaLabel => (
-              <PaneHeaderIconButton
-                id="clickable-edit-instance"
-                style={{ visibility: !instance ? 'hidden' : 'visible' }}
-                href={this.craftLayerUrl('edit', location)}
-                onClick={this.onClickEditInstance}
-                ariaLabel={ariaLabel}
-                icon="edit"
-              />
-            )}
-          </FormattedMessage>
-        </PaneMenu>
-      </IfPermission>
-    );
-
     if (!instance) {
       return (
         <Pane
@@ -455,7 +468,6 @@ class ViewInstance extends React.Component {
           defaultWidth={paneWidth}
           paneTitle={<FormattedMessage id="ui-inventory.editInstance" />}
           appIcon={<AppIcon app="inventory" iconKey="instance" />}
-          lastMenu={detailMenu}
           dismissible
           onClose={onClose}
         >
@@ -584,7 +596,7 @@ class ViewInstance extends React.Component {
       instanceStatusTerm: this.refLookup(referenceTables.instanceStatuses, get(instance, ['statusId'])).name || '-',
       instanceStatusCode: this.refLookup(referenceTables.instanceStatuses, get(instance, ['statusId'])).code || '-',
       instanceStatusSource: this.refLookup(referenceTables.instanceStatuses, get(instance, ['statusId'])).source || '-',
-      instanceStatusUpdatedDate: get(instance, ['statusUpdatedDate'], '-'),
+      instanceStatusUpdatedDate: instance?.statusUpdatedDate,
       modeOfIssuance: formatters.modesOfIssuanceFormatter(instance, referenceTables.modesOfIssuance) || '-',
       statisticalCodeIds: get(instance, ['statisticalCodeIds'], []),
     };
@@ -621,6 +633,9 @@ class ViewInstance extends React.Component {
 
     const contributors = get(instance, ['contributors'], []);
 
+    const natureOfContentTerms = get(instance, 'natureOfContentTermIds', [])
+      .map(termId => this.refLookup(referenceTables.natureOfContentTerms, termId).name);
+
     const descriptiveData = {
       publication: get(instance, ['publication'], []),
       editions: get(instance, ['editions'], []),
@@ -628,7 +643,7 @@ class ViewInstance extends React.Component {
       resourceTypeTerm: this.refLookup(referenceTables.instanceTypes, get(instance, ['instanceTypeId'])).name || '-',
       resourceTypeCode: this.refLookup(referenceTables.instanceTypes, get(instance, ['instanceTypeId'])).code || '-',
       resourceTypeSource: this.refLookup(referenceTables.instanceTypes, get(instance, ['instanceTypeId'])).source || '-',
-      natureOfContentTermIds: get(instance, ['natureOfContentTermIds'], []),
+      natureOfContentTerms: natureOfContentTerms || [],
       instanceFormatIds: get(instance, ['instanceFormatIds'], []),
       languages: get(instance, ['languages'], []),
       publicationFrequency: get(instance, ['publicationFrequency'], []),
@@ -697,18 +712,7 @@ class ViewInstance extends React.Component {
       acc10: !areAllFieldsEmpty(values(instanceRelationship)),
     };
 
-    const formattedStatusUpdatedDate = instanceData.instanceStatusUpdatedDate !== '-'
-      ? (
-        <Fragment>
-          <p>
-            <FormattedDate value={instanceData.instanceStatusUpdatedDate} />
-          </p>
-          <p>
-            <FormattedTime value={instanceData.instanceStatusUpdatedDate} />
-          </p>
-        </Fragment>
-      )
-      : noValue;
+    const formattedStatusUpdatedDate = getDateWithTime(instanceData.instanceStatusUpdatedDate);
 
     return (
       <Pane
@@ -727,7 +731,6 @@ class ViewInstance extends React.Component {
           </span>
         }
         paneSub={instanceSub()}
-        lastMenu={detailMenu}
         dismissible
         onClose={onClose}
         actionMenu={this.createActionMenuGetter(instance)}
@@ -855,6 +858,12 @@ class ViewInstance extends React.Component {
                   <KeyValue
                     label={<FormattedMessage id="ui-inventory.instanceStatusTerm" />}
                     value={checkIfElementIsEmpty(instanceData.instanceStatusTerm)}
+                    subValue={
+                      <FormattedMessage
+                        id="ui-inventory.item.status.statusUpdatedLabel"
+                        values={{ statusDate: formattedStatusUpdatedDate }}
+                      />
+                    }
                   />
                 </Col>
                 <Col xs={3}>
@@ -867,12 +876,6 @@ class ViewInstance extends React.Component {
                   <KeyValue
                     label={<FormattedMessage id="ui-inventory.instanceStatusSource" />}
                     value={checkIfElementIsEmpty(instanceData.instanceStatusSource)}
-                  />
-                </Col>
-                <Col xs={3}>
-                  <KeyValue
-                    label={<FormattedMessage id="ui-inventory.instanceStatusUpdatedDate" />}
-                    value={formattedStatusUpdatedDate}
                   />
                 </Col>
               </Row>
@@ -1164,7 +1167,10 @@ class ViewInstance extends React.Component {
               </Row>
 
               <Row>
-                <Col xs={3}>
+                <Col
+                  data-test-nature-of-content-terms
+                  xs={3}
+                >
                   <KeyValue
                     label={<FormattedMessage id="ui-inventory.natureOfContentTerms" />}
                     value={convertArrayToBlocks(descriptiveData.natureOfContentTermIds)}