+++ src/components/InstancePlugin/InstancePlugin.js
@@ -4,17 +4,21 @@ import { FormattedMessage } from 'react-intl';
 
 import { Pluggable } from '@folio/stripes/core';
 
+import css from './InstancePlugin.css';
+
 const InstancePlugin = ({ onSelect }) => (
-  <Pluggable
-    aria-haspopup="true"
-    dataKey="instances"
-    searchButtonStyle="default"
-    searchLabel="+"
-    selectInstance={onSelect}
-    type="find-instance"
-  >
-    <span><FormattedMessage id="ui-inventory.findInstancePluginNotFound" /></span>
-  </Pluggable>
+  <div className={css.marginTop}>
+    <Pluggable
+      aria-haspopup="true"
+      dataKey="instances"
+      searchButtonStyle="default"
+      searchLabel="+"
+      selectInstance={onSelect}
+      type="find-instance"
+    >
+      <span><FormattedMessage id="ui-inventory.findInstancePluginNotFound" /></span>
+    </Pluggable>
+  </div>
 );
 
 InstancePlugin.propTypes = {