+++ test/bigtest/interactors/inventory.js
@@ -11,11 +11,14 @@ import {
   Interactor,
 } from '@bigtest/interactor';
 
+import DateRangeFilterInteractor from '@folio/stripes-smart-components/lib/SearchAndSort/components/DateRangeFilter/tests/interactor';
+
 import {
   MultiSelectFilterInteractor,
 } from './filters';
 
 @interactor class InventoryHeaderDropdownMenu {
+  clickNewInstanceBtn = clickable('#clickable-newinventory');
   clickItemsInTransitReportBtn = clickable('#dropdown-clickable-get-report');
   itemsInTransitReportBtnIsVisible = isVisible('#dropdown-clickable-get-report');
   clickSaveInstancesUIIDsBtn = clickable('#dropdown-clickable-get-items-uiids');
@@ -32,6 +35,14 @@ import {
   exportInstancesJSONBtnIsVisible = isVisible('#dropdown-clickable-export-json');
   isExportInstancesJSONBtnDisabled = property('#dropdown-clickable-export-json', 'disabled');
   isExportInstancesJSONIconPresent = isPresent('#dropdown-clickable-export-json [class*=icon-download]');
+}
+
+@interactor class CreatedDateFilterInteractor {
+  static defaultScope = '#createdDate';
+
+  clear = clickable('button[icon="times-circle-solid"]');
+  open = clickable('[id*="accordion-toggle-button-"]');
+  filter = new DateRangeFilterInteractor();
 }
 
 export default @interactor class InventoryInteractor {
@@ -69,6 +80,7 @@ export default @interactor class InventoryInteractor {
   resourceTypeFilter = scoped('#resource', MultiSelectFilterInteractor);
   formatFilter = scoped('#format', MultiSelectFilterInteractor);
   modeFilter = scoped('#mode', MultiSelectFilterInteractor);
+  createdDate = new CreatedDateFilterInteractor();
 
   whenLoaded() {
     return this.when(() => this.isLoaded);