+++ test/bigtest/interactors/item-view-page.js
@@ -18,12 +18,15 @@ import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumn
 }
 
 @interactor class HeaderDropdownMenu {
+  hasEdit = isPresent('[data-test-inventory-edit-item-action]');
   clickEdit = clickable('[data-test-inventory-edit-item-action]');
   hasDuplicate = isPresent('[data-test-inventory-duplicate-item-action]');
   clickDuplicate = clickable('[data-test-inventory-duplicate-item-action]');
   hasNewRequestItem = isPresent('[data-test-inventory-create-request-action]');
   hasMarkAsMissing = isPresent('[data-test-mark-as-missing-item]');
   clickMarkAsMissing = clickable('[data-test-mark-as-missing-item]');
+  hasMarkAsWithdrawn = isPresent('[data-test-mark-as-withdrawn-item]');
+  clickMarkAsWithdrawn = clickable('[data-test-mark-as-withdrawn-item]')
   hasDeleteItem = isPresent('[data-test-inventory-delete-item-action]');
   clickDelete = clickable('[data-test-inventory-delete-item-action]');
 }
@@ -41,8 +44,8 @@ import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumn
   hasDeleteModal = isPresent('#confirmDeleteItemModal');
   hasMarkAsMissingModal = isPresent('[data-test-missingConfirmation-modal]');
   markAsMissingModal = scoped('[data-test-missingConfirmation-modal]');
-  hasEditItemButton = isPresent('[data-test-clickable-edit-item]');
-  clickEditItemButton = clickable('[data-test-clickable-edit-item]');
+  hasMarkAsWithdrawnModal = isPresent('[data-test-withdrawn-confirmation-modal]');
+  markAsWithdrawnModal = scoped('[data-test-withdrawn-confirmation-modal]');
   cannotDeleteItemModal = new ModalInteractor('[data-test-cannot-delete-item-modal]');
   cannotDeleteItemModalBackButton = new ButtonInteractor('[data-test-cannot-delete-item-back-action]');
   confirmDeleteItemModal = new ConfirmationModalInteractor('#confirmDeleteItemModal');