+++ test/bigtest/tests/item-view-page-test.js
@@ -32,10 +32,6 @@ describe('ItemViewPage', () => {
         expect(ItemViewPage.title).to.equal(`Item record ${item.barcode} ${item.status.name}`);
       });
 
-      it('displays the edit button in the pane header', () => {
-        expect(ItemViewPage.hasEditItemButton).to.be.true;
-      });
-
       it('displays the "Collapse all" button', () => {
         expect(ItemViewPage.collapseAllButton.isPresent).to.be.true;
       });
@@ -75,19 +71,13 @@ describe('ItemViewPage', () => {
         });
       });
 
-      describe('clicking on edit button in the pane header', () => {
-        beforeEach(async () => {
-          await ItemViewPage.clickEditItemButton();
-        });
-
-        it('should redirect to item edit page', () => {
-          expect(ItemEditPage.$root).to.exist;
-        });
-      });
-
       describe('pane header dropdown menu', () => {
         beforeEach(async () => {
           await ItemViewPage.headerDropdown.click();
+        });
+
+        it('should show an edit menu item', () => {
+          expect(ItemViewPage.headerDropdownMenu.hasEdit).to.be.true;
         });
 
         it('should show a duplicate menu item', () => {
@@ -96,6 +86,10 @@ describe('ItemViewPage', () => {
 
         it('should show a mark as missing item', () => {
           expect(ItemViewPage.headerDropdownMenu.hasMarkAsMissing).to.be.true;
+        });
+
+        it('should show a mark as withdrawn item', () => {
+          expect(ItemViewPage.headerDropdownMenu.hasMarkAsWithdrawn).to.be.true;
         });
 
         it('should show a delete menu item', () => {
@@ -149,6 +143,31 @@ describe('ItemViewPage', () => {
             });
           });
         });
+
+        describe('clicking on mark as withdrawn', () => {
+          beforeEach(async () => {
+            await ItemViewPage.headerDropdownMenu.clickMarkAsWithdrawn();
+          });
+
+          it('should open a withdrawn confirmation modal', () => {
+            expect(ItemViewPage.hasMarkAsWithdrawnModal).to.exist;
+          });
+
+          describe('clicking on "Confirm" button', () => {
+            beforeEach(async () => {
+              await ItemViewPage.confirmButton.click();
+            });
+
+            it('should close a withdrawn confirmation modal', () => {
+              expect(ItemViewPage.hasMarkAsWithdrawnModal).to.be.false;
+            });
+
+            it('should change item status to "Withdrawn"', () => {
+              expect(ItemViewPage.loanAccordion.keyValues(2).text).to.be.equal('Withdrawn');
+            });
+          });
+        });
+
         describe('clicking on delete', () => {
           beforeEach(async () => {
             await ItemViewPage.headerDropdownMenu.clickDelete();
@@ -243,10 +262,6 @@ describe('ItemViewPage', () => {
         await ItemViewPage.whenLoaded();
       });
 
-      it('displays the edit button in the pane header', () => {
-        expect(ItemViewPage.hasEditItemButton).to.be.false;
-      });
-
       it('should show a dropdown menu in the pane header', () => {
         expect(ItemViewPage.hasHeaderDropdown).to.be.undefined;
       });
@@ -267,8 +282,14 @@ describe('ItemViewPage', () => {
         expect(ItemViewPage.headerDropdownMenu.hasNewRequestItem).to.be.false;
       });
 
-      it('displays the edit button in the pane header', () => {
-        expect(ItemViewPage.hasEditItemButton).to.be.false;
+      describe('pane header dropdown menu', () => {
+        beforeEach(async () => {
+          await ItemViewPage.headerDropdown.click();
+        });
+
+        it('should not display an edit item', () => {
+          expect(ItemViewPage.headerDropdownMenu.hasEdit).to.be.false;
+        });
       });
     });
   });