+++ test/ui-testing/new-title.js
@@ -33,6 +33,8 @@ module.exports.test = function uiTest(uiTestCtx) {
 
       it('should create new instance ', (done) => {
         nightmare
+          .wait('[data-test-inventory-instances] #paneHeaderpane-results [data-test-pane-header-actions-button]')
+          .click('[data-test-inventory-instances] #paneHeaderpane-results [data-test-pane-header-actions-button]')
           .wait('#clickable-newinventory')
           .click('#clickable-newinventory')
           .wait(55)
@@ -106,8 +108,8 @@ module.exports.test = function uiTest(uiTestCtx) {
           .wait('input[name="notes[2].note"]')
           .insert('input[name="notes[2].note"]', notes[2])
           .wait(55)
-          .wait('#clickable-create-instance')
-          .click('#clickable-create-instance')
+          .wait('#clickable-save-instance')
+          .click('#clickable-save-instance')
           .wait('#clickable-new-holdings-record')
           .then(() => { done(); })
           .catch(done);